Restructure for proper routing
