+++ app/__init__.py
@@ -8,9 +8,10 @@ db = SQLAlchemy()
 
 def create_app(config_name):
     app = Flask(__name__, instance_relative_config=True)
-    app.config.from_object(app_config['testing'])
-    # app.config.from_pyfile('config.py')
-    # app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+    app.config.from_object(app_config[config_name])
+    app.config.from_pyfile('config.py')
     db.init_app(app)
     
+    from app.apis import apiv1_bp
+    app.register_blueprint(apiv1_bp , url_prefix='/apiv1')
     return app
