+++ app/apis/recipie.py
@@ -1,10 +1,21 @@
 from flask import request
-from flask_restplus import Resource, Namespace
+from flask_restplus import Resource, Namespace, fields
 
-# from app.apis.models.recipie import Recipie
-from ..functionality.serializers import recipe
+from app.models.recipie import Recipie
+# from ..functionality.serializers import recipe
 
 api = Namespace('recipie', description='Recipie related functionality')
+
+recipe = api.model('recipie', {
+    'recipie_id' : fields.Integer(readOnly=True, description='recipie unique identifier'),
+    'recipie_name' : fields.String(required=True, description='recipie name'),
+    'ingedients' : fields.String(required=True, description='A description of the ingredients to compile the recipie'),
+    'created_by' : fields.Integer(readOnly=True, description='Which User created this nanka'),
+    'attached_category' : fields.Integer(readOnly=True, description='Which category does this recipe belong to'),
+    'date_created' : fields.DateTime(readOnly=True, description='Date Created'),
+    'date_modified' : fields.DateTime(readOnly=True, description='Date modified')
+})
+
 
 @api.route('/')
 class RecipieCollection(Resource):
